Qazaq Cyrillic & Latine Update
The implementation is done for Qazaq words written in latine:

New data of Qazaq words and themes are added
The functions that help to convert Cyrillic letters to Latine and vice versa are added. Cyrillic words are also able to be converted to Latine. These functions allow the Reinforcement learning algorithm to work with Latine and Cyrillic words.
All previous algorithms that worked with English words now work with Qazaq Cyrillic (database) and Latine (user's written words and converted words from database) words.
The additional function that compares two words (user's and the one that is given to a user) is added. This function does comparison and can find mistakes if:
the order of two letters is reversed, one letter is missed, the extra letter is added, and/or one letter is misspelled. It can find several mistakes in one instance. For example, if the word which is given to a user is "әлемдік" and a user wrote "äelmık", the function will provide two mistakes: reversed mistake of ["l", "e"] and will show that one letter is missed ["d"].